requirements.txt
beautifulsoup4>=4.9.3
requests>=2.25.1
python-dotenv>=0.19.0
Flask>=2.0.1
PyGithub>=1.55
Pillow>=8.3.1
python-magic>=0.4.24
python-jose>=3.3.0
python-multipart>=0.0.5
gunicorn>=20.1.0
